Relax reservation purge due to any invalid uid after creation time.
Previously, _build_uid_list() would return ESLURM_USER_ID_MISSING if any
uid could not be retrieved off the requested user string through
uid_from_string() and the callers would purge the reservation if already
created (or not proceed to create it at creation time). Eventualities like
a user removed from LDAP could trigger this, even if the rest of the
reservation users uids could be validated.

This commit adds a new "strict" bool parameter which is only set to true
at creation time preserving the previous behavior. After creation time
it is set to false making it so that if an invalid user is found,
instead of early bail with the error, continue checking the rest of the
users and build the uid list off the valid ones, even if the count
doesn't match that of the requested users string. If the list has zero
valid uids, then also return the error.

Bug 11483