Don't use slurm_verify_cpu_bind to copy cpu_bind_type
Dedicated block in slurm_verify_cpu_bind for the call with a NULL arg
was an overcomplicated way to to set cpu_bind_type to a default value if
nothing else was already set.

Special handling applied to CPU_BIND_VERBOSE. If it's set in defaults we
add it to user specified --cpu-bind.

Bug 11247